Make Arguments an implementation detail of db::Context
